implementing integration test structure
